@@ -1,13 +1,13 @@
 {
   "name": "md2bb",
-  "version": "0.1.0",
+  "version": "0.1.1",
   "description": "steam-bbcode written by markdown like syntax",
   "main": "./index.js",
   "scripts": {
-    "test": "./node_modules/.bin/mocha --compilers js:babel-register test/md2bb/*.js",
-    "build": "./node_modules/.bin/webpack -p",
+    "test": "mocha --compilers js:babel-register test/md2bb/*.js",
+    "build": "webpack",
     "start": "npm run build && npm run test",
-    "watch": "./node_modules/.bin/webpack --watch"
+    "watch": "webpack --watch"
   },
   "author": "tenmihi <tenmihi@gmail.com>",
   "license": "MIT",
@@ -24,7 +24,9 @@
     "eslint-plugin-jsx-a11y": "^5.0.3",
     "eslint-plugin-react": "^7.0.1",
     "mocha": "^3.4.2",
-    "webpack": "^2.6.1",
-    "webpack-node-externals": "^1.6.0"
+    "webpack": "^3.5.5"
+  },
+  "dependencies": {
+    "simple-parser": "tenmihi/simple-parser#master"
   }
 }
